Verify topology in MS
